Fix issue #941

+++ sm-core/src/main/java/com/salesmanager/core/business/repositories/content/PageContentRepository.java
@@ -9,7 +9,7 @@ import com.salesmanager.core.model.content.Content;
 import com.salesmanager.core.model.content.ContentType;
 import com.salesmanager.core.model.merchant.MerchantStore;
 
-public interface PageContentRepository extends PagingAndSortingRepository<MerchantStore, Long> {
+public interface PageContentRepository extends PagingAndSortingRepository<MerchantStore, Integer> {
